🚚 refactor: rename featureFlag store to serverConfig store
