+++ src/app/chat/(desktop)/features/ChatHeader/HeaderAction.tsx
@@ -2,9 +2,9 @@ import { ActionIcon } from '@lobehub/ui';
 import { PanelRightClose, PanelRightOpen } from 'lucide-react';
 import { memo } from 'react';
 import { useTranslation } from 'react-i18next';
+import { featureFlagsSelectors, useServerConfigStore } from 'src/store/serverConfig';
 
 import { DESKTOP_HEADER_ICON_SIZE } from '@/const/layoutTokens';
-import { featureFlagsSelectors, useFeatureFlagStore } from '@/store/featureFlags';
 import { useGlobalStore } from '@/store/global';
 
 import SettingButton from '../../../features/SettingButton';
@@ -18,7 +18,7 @@ const HeaderAction = memo(() => {
     s.toggleChatSideBar,
   ]);
 
-  const { isAgentEditable } = useFeatureFlagStore(featureFlagsSelectors);
+  const { isAgentEditable } = useServerConfigStore(featureFlagsSelectors);
 
   return (
     <>
