+++ src/features/AgentSetting/AgentPlugin/index.tsx
@@ -5,11 +5,11 @@ import { LucideToyBrick, LucideTrash2, Store } from 'lucide-react';
 import { memo, useState } from 'react';
 import { Trans, useTranslation } from 'react-i18next';
 import { Center, Flexbox } from 'react-layout-kit';
+import { featureFlagsSelectors, useServerConfigStore } from 'src/store/serverConfig';
 
 import { FORM_STYLE } from '@/const/layoutTokens';
 import PluginStore from '@/features/PluginStore';
 import PluginTag from '@/features/PluginStore/PluginItem/PluginTag';
-import { featureFlagsSelectors, useFeatureFlagStore } from '@/store/featureFlags';
 import { pluginHelpers, useToolStore } from '@/store/tool';
 import { toolSelectors } from '@/store/tool/selectors';
 
@@ -29,7 +29,7 @@ const AgentPlugin = memo(() => {
     s.toggleAgentPlugin,
   ]);
 
-  const { showDalle } = useFeatureFlagStore(featureFlagsSelectors);
+  const { showDalle } = useServerConfigStore(featureFlagsSelectors);
   const installedPlugins = useToolStore(toolSelectors.metaList(showDalle), isEqual);
   const useFetchInstalledPlugins = useToolStore((s) => s.useFetchInstalledPlugins);
