+++ src/layout/GlobalProvider/index.tsx
@@ -10,7 +10,7 @@ import {
   LOBE_THEME_NEUTRAL_COLOR,
   LOBE_THEME_PRIMARY_COLOR,
 } from '@/const/theme';
-import { FeatureFlagStoreProvider } from '@/store/featureFlags';
+import { ServerConfigStoreProvider } from '@/store/serverConfig';
 import { getAntdLocale } from '@/utils/locale';
 
 import AppTheme from './AppTheme';
@@ -55,9 +55,9 @@ const GlobalLayout = async ({ children }: GlobalLayoutProps) => {
           defaultPrimaryColor={primaryColor?.value as any}
         >
           <StoreInitialization />
-          <FeatureFlagStoreProvider featureFlags={serverFeatureFlags}>
+          <ServerConfigStoreProvider featureFlags={serverFeatureFlags}>
             {children}
-          </FeatureFlagStoreProvider>
+          </ServerConfigStoreProvider>
           <DebugUI />
         </AppTheme>
       </Locale>
